fix: 评价无法上传评分

+++ jsoup-2030/src/main/java/dam/jsoup/updatereport/updatreport/vo/GoodList.java
@@ -29,7 +29,7 @@ public class GoodList {
     /**
      * 评分
      */
-    private Integer rate;
+    private BigDecimal rate;
     /**
      * 创造用户名
      */
